extract NodeStatusTuple inner class
